Add context routing
