+++ mikado-app/package.json
@@ -10,6 +10,8 @@
     "firebase": "^9.17.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-google-button": "^0.7.2",
+    "react-router-dom": "^6.8.1",
     "react-scripts": "5.0.1",
     "reactflow": "^11.5.5",
     "web-vitals": "^2.1.4"
